recommendation/pyproject.toml
[project]
name = "AnimmendationRecommender"
version = "1.0.0"
description = ""
authors = [
  { name = "Henrique Tavares Aguiar", email = "ricktavares76@gmail.com" },
]
dependencies = [
    "river>=0.14.0",
    "grpcio>=1.50.0",
    "grpcio-tools>=1.50.0",
    "prisma>=0.7.1",
    "mpire>=2.6.0",
    "annoy>=1.17.1",
    "scikit-learn>=1.1.3",
    "loguru>=0.6.0",
    "more-itertools>=9.0.0",
    "numba>=0.56.4",
    "pysftp==0.2.8",
    "python-dotenv>=0.21.0",
]
requires-python = ">=3.10"
license = { text = "MIT" }
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "notebook>=6.5.2",
    "black>=22.10.0",
    "flake8>=5.0.4",
    "progress>=1.6",
    "mypy-protobuf>=3.4.0",
    "ipywidgets>=8.0.2",
]

[tool.pdm.scripts]
start = { shell = "python src/server.py" }
grpc = { cmd = [
  "python",
  "-m",
  "grpc_tools.protoc",
  "--plugin=__pypackages__/3.10/bin/protoc-gen-mypy",
  "-I=src/infra/grpc/",
  "--python_out=src/infra/grpc/",
  "--grpc_python_out=src/infra/grpc/",
  "--mypy_out=src/infra/grpc/",
  "src/infra/grpc/recommender.proto",
] }
